fix(amfd): the emergency context was keyed by a truncated u32 AMF UE NGAP ID (#353) (#355)
Closes #353 -- all three of its fix bullets.

The AMF UE NGAP ID is INTEGER (0..2^40-1) (TS 38.413 §9.3.3.1) and the rest of
amfd carries it as u64, but EmergencyHandler keyed active_contexts on u32, so the
call site #72 added truncated with `as u32`. Two UEs whose ids differ only above
bit 32 shared one entry: HashMap::insert means the second OVERWRITES the first,
losing whether it was authenticated and which PDU session it held; either UE's
release_emergency frees the other's context; and active_count() under-reports,
which is the number an operator watches during an emergency. Latent rather than
live -- amfd allocates ids from a low counter -- but a regulatory-grade failure
mode on an emergency-services path.

Widened the map and all three method parameters to u64, and deleted the cast plus
the comment that documented it.

ALSO ADDED emergency_context(id) -> Option<&EmergencyContext>, because without it
the stored context is WRITE-ONLY: assign_emergency_pdu_session records a PDU
session id nothing outside the type could read, so a context silently replaced by
a colliding key was undetectable from the outside -- the same blindness that let
the truncation sit unnoticed. It is what lets the test assert content rather than
only a count.

The test uses 1 and 1 + (1 << 32), both legal 40-bit ids, and asserts up front
that they DO collide under the old key so the fixture cannot silently stop
exercising the defect. It registers them with opposite has_supi and different PDU
sessions, so an overwrite surfaces as the wrong authentication state rather than
only as a smaller count, then releases one and asserts the other still releases --
reachable only if two entries exist. Revert-verified: re-narrowing the map to u32
with `as u32` at each acquisition fails it at the active_count() == 2 assertion.

SCOPE NOTE: assign_emergency_pdu_session and release_emergency still have no
production caller (only handle_emergency_registration got one, in #72), so
emergency contexts LEAK -- nothing releases one when the UE context goes away and
active_count() only ever grows. That is a separate defect from the key width and
is filed rather than widened in here.

Workspace 6582 tests 0 failures, clippy 0 errors, fmt clean.

Spec: specs/fix-amfd-emergency-context-u64-key.md

+++ src/bins/nextgcore-amfd/src/ngap_path.rs
@@ -1988,13 +1988,9 @@ impl NgapServer {
                 .and_then(|s| s.amf_ue.supi.clone())
                 .is_some()
                 || req.suci.is_some();
-            // The handler is keyed by a u32; the AMF UE NGAP ID is 40 bits on the wire
-            // (TS 38.413), so the truncation is recorded rather than hidden. It is safe
-            // here because the key only has to be unique among CONCURRENT emergency
-            // registrations, and #353 tracks widening it.
             let ctx = self
                 .emergency
-                .handle_emergency_registration(amf_ue_ngap_id as u32, has_supi);
+                .handle_emergency_registration(amf_ue_ngap_id, has_supi);
             log::warn!(
                 "EMERGENCY registration from UE {amf_ue_ngap_id}: authenticated={}, \
                  reg_type={:?}, emergency DNN '{}' (TS 24.501 §5.5.1.2, TS 23.167)",
